+++ variables.tf
@@ -59,9 +59,16 @@ variable "extra_tags" {
 }
 
 variable "georeplication_locations" {
-  description = "A list of Azure locations where the container registry should be geo-replicated."
-  type        = list(string)
-  default     = null
+  description = <<DESC
+  A list of Azure locations where the container registry should be geo-replicated. Only activated on Premium SKU.
+  Supported properties are:
+    location                = string
+    zone_redundancy_enabled = bool
+    tags                    = map(string)
+  or this can be a list of `string` (each element is a location)
+DESC
+  type        = list(any)
+  default     = []
 }
 
 variable "images_retention_enabled" {
@@ -74,29 +81,6 @@ variable "images_retention_days" {
   description = "Specifies the number of images retention days."
   type        = number
   default     = 90
-}
-
-variable "logs_destinations_ids" {
-  type        = list(string)
-  description = "List of destination resources Ids for logs diagnostics destination. Can be Storage Account, Log Analytics Workspace and Event Hub. No more than one of each can be set. Empty list to disable logging."
-}
-
-variable "logs_categories" {
-  type        = list(string)
-  description = "Log categories to send to destinations."
-  default     = null
-}
-
-variable "logs_metrics_categories" {
-  type        = list(string)
-  description = "Metrics categories to send to destinations."
-  default     = null
-}
-
-variable "logs_retention_days" {
-  type        = number
-  description = "Number of days to keep logs on storage account"
-  default     = 30
 }